+++ internal/emulation/qbittorrent.go
