+++ internal/emulation/qbittorrent_test.go
